renamed a file for more consistent naming scheme
